CastDetailsView: Create MediaRouteControllerContentManager
Create a content manager to handle the UI logic within the media route
controller. This class is currently being used by
MediaRouteControllerDialog, and will be used by the cast details view in
the future.

This CL follows the refactor mentioned in go/cast-details-view.

Bug: b/395947047
Flag: NONE refactor
Test: Cast to a device, click on the Cast tile, launched the dialog and
verified that the controller works.

Change-Id: I80bbd658d7c2686c4cd1f4c3eafd74eaeeb002da