Cleanup ShadeDisplayAware dagger modules
shadeDisplaysInteractor was bound in SystemUIModule (that is available for all sysui variants), despite using NotificationStackRebindingHider, ShadeExpandedStateInteractor and WindowRootView (that were only provided in ShadeModule, not available in SysUI variants other than Google and aosp).

The issue was mistakenly fixed using optional bindings. This cl properly fixes it by binding it only in ShadeModule.

It is **not** possible to bind the entire ShadeDisplayAwareModule in the same way (as an include of ShadeModule, for example) as there are several classes in SystemUIModule that need the objects provided by ShadeDisplayAwareModule.

Bug: 362719719
Bug: 396159919
Test: Builds
Flag: com.android.systemui.shade_window_goes_around
Change-Id: If9d9e476c50013501fbabd41234012e31304e401