Add test for KeyguardController removing AOD during unlock
There is a heavily-exercised special case for KeyguardController:
if the keyguard is going away but AOD is still showing, calling
setKeyguardShown(true, false) will be treated as a request to
remove the AOD bit without interfering with the keyguardGoingAway
state.

This is a surprising special case that was not caught by unit tests
when a change to the surrounding code was submitted. Adding a test
codifies that this is a deliberate "feature" and shouldn't be
changed without a plan for alternative ways of clearing the AOD
status during keyguardGoingAway

Change-Id: Ide11cc6cfd33e826c54da704fd66997938735cc3
Flag: EXEMPT test only
Test: atest 'DisplayContentTests#testKeyguardGoingAwayWhileAodShown'
Bug: 396134691