Fixes the regression that the DesktopLaunchParams not working
Avoid modify the tasks that are managed by a created-by-organizer Task.
This is just a short-term solution that avoid the regression.
In general, the Desktop launch-params modifier should not calculate
the Task bounds for the Tasks that are not going to show in Desktop.

This CL also avoid calculate task bounds when no target bounds requested
and the Task already has override bounds.

Bug: 396108436
BUg: 396659743
Test: StartActivityTests#testStartActivityWithLaunchBounds
Test: ActivityLifecycleTopResumedStateTests#testTopPositionSwitchBetweenVisibleActivities
Flag: EXEMPT(bug fix)
Change-Id: Ib510d231c0a6be1fb3975e7d61c92117d16fb8ec