[sb] update wifi icons
This version of the new wifi icons only defines 3 bars for level. Since
the definitions of the icons in WifiUtils and WifiIcons both use an
array to index from level -> icon, we can just re-use the same
value for levels 2 and 3, and define level 4 as a full-3-bar icon.

Test: manual
Test: SystemIconsScreenshotTest
Bug: 391606042
Flag: com.android.settingslib.flags.new_status_bar_icons

Change-Id: I794ffb3a7e515560ab446f33146d5fb8f2f92b85