[AnimLib] Reuse same unique token for same view
This change enables clients of ViewTransitionRegistry to reuse same
uniue token associated with a particular view (till it is attached to a
window). This optimization reduces the overhead of generating different
unique tokens mapped to the same view and ensures that there is a single
token, entry and listener associated with a view in the registry (this
detail is completely abstracted out from clients).

Bug: 393241010
Flag: com.android.systemui.decouple_view_controller_in_animlib
Test: GhostedViewTransitionAnimatorControllerTest, ViewTransitionRegistryTest
Change-Id: If0dbd9de5809068476cfe69268b17429f5dda11d