Don't set bubble launch flags on non-chat shortcut bubbles
Previously the isShortcutBubble path was only for chat bubbles using
shortcuts, however, with bubble anything we allow bubbling shortcuts
on launcher. These were still being tagged as "launch from bubble"
when they shouldn't be.

This CL only sets those flags if it's a chat bubble. If it's a
shortcut bubble from bubble anything, only set the multiple task flag
which is consistent with intent/pendingIntent app bubbles.

This CL also adds a bunch of verification in the BubbleTaskViewListener
test for the flags set on intents / activity options.

It seems like there are some issues with robolectric and PendingIntents.
I couldn't figure out a way to capture PendingIntents in robolectric, so
I modified the code to always use a fillInIntent & the flags are checked
on that instead.

Flag: EXEMPT bug fix
Test: atest BubbleTaskViewListenerTest
Bug: 395145961
Bug: 395178844
Change-Id: Ia01f740ab80d46f8b6533d9f87ba319a259bd5e0