Make CreateUserDialogController fully responsible for dialog dismissal
In ag/30936343 dismissal of the dialog was moved from CreateUserDialogController to CreateUserActivity, which lead to potential NPE if the controller is used directly without being hosted within the activity. This change brings dismissal of the dialog back into CreateUserDialogController and makes it fully responsible for the dismissal of the dialog.

Bug: 398042304
Bug: 398767163
Test: atest UserSettingsTest && atest CreateUserActivityTest
Flag: android.multiuser.place_add_user_dialog_within_activity
Change-Id: I463fe89f09bfb6cd0866b37251c6cda6e2193075