Fix displaying advanced protection dialog on top of the Wi-Fi quick tile
Fixes an issue where the advanced protection dialog was missing when
connecting to a WEP network via the Wi-Fi quick tile with advanced
protection enabled.

The fix involves making the advanced protection Intent explicit to the
Settings package, and adjusting the window type of the dialog to show on
top of the WEP quick tile, similar to the WEP dialog.

Bug: 389081238
Test: Visual test
Flag: EXEMPT bug fix
Change-Id: Ie6f60c88d65e08e4a0b51bfa04a179fa965b27b1