Fix race condition in testExpansionAndCollapse
The request to expand the stack takes a runnable that is called
when the expand animation ends. But if more bubbles are added
during the animation, then a new animation will start. This change
adds another call to make sure we wait for any additional animation
that may be running before asserting on the positions of the bubbles.

Change-Id: Ic69ed6a18ecd3e135a25d3e544d060fddc0e8279
Flag: EXEMPT test only
Bug: 397843857
Test: monitor flake rate