Revert "Cleanup flag for notification sound Uri permission check"
This reverts commit f73aba3ee40b5ec255ef6fa408b360050718968f.

Reason for revert: CVE-2025-22420 A-337775777 is being removed from ASB#2025-04 due to an possible app compatibility regression. Partners are recommended to remove this fix and should not reference it in any public documentation. An updated fix will be required as part of a future Android Security Bulletin.

Pixel Impact: Internal testing found WhatsApp java crashes with this fix. Detected on Dogfood builds. We recommend Pixel removal if possible. Crash report here b/390020418

Change-Id: I18ab82279e67ad0b7b8de509b23e1b7607b1dbbe