[Notif redesign] Remove action_list_margin_target
This special ID for adding an extra margin for actions is no longer
needed. The special margin we added for smart replies is also not
needed anymore.

Instead, we're keeping the actions_container visible to act like a
margin for the notification when there are no actions.

Bug: 378660052
Test: manual + screenshot tests
Flag: android.app.notifications_redesign_templates
Change-Id: If434b7cde20102ad0740681aaf94450324dd801e