[Notif redesign] Use layout_weight instad of custom FrameLayout
This was previously done in I775a0bb29944922581d01f237dbbc5580ced00c4
for the other layouts, but not for the legacy messaging style. And from
there, it also ended up being forked in the new conversation/call
layouts for the redesign, so let's remove it to avoid some issues with
height.

Bug: 378660052
Bug: 401282728
Test: screenshot tests should be unchanged
Flag: android.app.notifications_redesign_templates
Change-Id: Ibfbf9071c03a2fa86bcddae44c28122164ff10e6