Implementing magnetic attachment using springs.
The ability to re-attach a magnetic notification to its neighbors is
being implemented. After being detached, crossing a new attachment
threshold reactivates the magnetic pulling. To avoid jump artificats,
the MagneticRowListeners decide if running animators must be cancelled
when setting a new translation if we wish to track the input gesture
directly in an eager fashion. This happens if the drag gestures are
larger than the set touch slop.

Test: MagneticNotificationRowManagerImplTest
Flag: com.android.systemui.magnetic_notification_swipes
Bug: 397418247
Change-Id: I04347232b4d7f386efc9c489184cf24354540f14