[flexiglass] Don't expand other dual shade when swiping in this one
Bug was that swiping down, on the other half of an overlay would expand
the other overlay. This was because the user actions were bound to the
left/right half. Instead, they should be bound to the right/left half of
the top edge.

Fix: 402375503
Test: verified manually by swiping down from the status bar or from the
middle, on either side (left/right) in both gone and lockscreen with or
without either overlay being expanded
Flag: com.android.systemui.scene_container

Change-Id: I6a08f49817a69fba52707a23c3fe3fa48cee63a1