Simplify slider animation code and reduce spring stiffness
That code was there to address that the Volume Dialog and Volume Slider
handle immediate volume state differently. Volume Dialog used to lack
any immediate state and relied on the Slider to hold it for a smoother
touch experience. Not it has this state in the View Model which allows
me to simplify the Composable and make a smoother animation when the
user holds a physical button

Flag: com.android.systemui.volume_redesign
Fixes: 398303967
Test: atest VolumeDialogScreenshotTest
Test: atest VolumePanelScreenshotTest
Test: manual on the foldable. Adjust volume with touch and physical
buttons

Change-Id: I80957ed57b08096c0be40f63c19d79be6246243d